Add .vscode/ to .gitignore
We really have the choice to either check in a standard set of vscode's
files (if that's something the team cares about) - or ignore them so
that it's not noisy for people that use the repo with VS Code.

The former requires more discussion and effort, so ignoring `.vscode/`
for now makes the most sense.