Workaround .NET SDK bug with pipelines package
This is a workaround for https://github.com/dotnet/sdk/issues/14019
which prevents the current prerelease dotnet SDK
`5.0.100-rc.1.20452.10` from successfully building the repo.

Users attempting to build with the bugged release will see a compiler
error related to the `System.IO.Pipelines` assembly.

The explicit reference to `System.IO.Pipelines` is harmless long term,
but can be removed when the underlying bug is fixed and shipped in a GA
release (likely 5.0.100 GA).