Remove automatic protobuf conversion (#557)
Fixes: #442

We're making it absolutely clear which methods do JSON (most of them)
and which do gRPC. We have the ability to do both 'raw' functionality
and optimized paths for JSON+HTTP and gRPC+Protobuf for invoke. We're
not going to add 'raw' or Protobuf support for other building blocks
until there's a specific use case.

For the most part the functionality that's being removed here isn't used
anymore since that was part of the invoke changes.