Improve actor getting started and add more docs
Fixes: #546
Fixes: #612

This change runs through the getting started docs and improves trims all
of the unnecessary fat. I've reorganized all of the details that aren't
relevant as part of the first tutorial into separate docs.

I've also fleshed out all of the content we have to be relevant to a new
user and added new sections like DI and logging.