docs fix: only UseHttpsRedirection for non-dev env
Moved the UseHttpsRedirection to the else block of the IsDevelopment
check, and add a comment describing why.

The Dapr runtime reports and error and aborts the request after an HTTP
redirection. By default ASP.Net Core uses port 5000 for HTTP, and port
5001 for HTTPs. This means that if the instructions are followed exactly
the example won't work. The VSCode extension also breaks.