Fix issue where state did not reset on actor error
When an actor call fails, it is imperative that the state is reset.
In the remoting path, we were catching the exception and returning
an error response instead of propagating the exception. Without this,
the error path was never triggered. This allowed for state from
failed requests to persist on subsequent requests.

https://github.com/dapr/dotnet-sdk/issues/762

Signed-off-by: Hal Spang <halspang@microsoft.com>