Add support for Query API
This commit adds methods to the DaprClient to call the new Query API.
Currently, a raw query string is provided to the API. The API is
still in the alpha stage which may cause shifts in how the API works.
This should not effect how the SDK exposes the call, but may change
the underlying semantics.

https://github.com/dapr/dotnet-sdk/issues/777

Signed-off-by: Hal Spang <halspang@microsoft.com>