Fix: Duration-based scheduling is properly formed (#1462)
Fixed: Duration-based scheduling should be properly formatted

When using a duration-based schedule value (e.g. "5s", it's important that the value be preceded by "@every " so it's properly evaluated by the runtime. While a value of "5s" is perfectly valid in the `dueTime` property, the scheduler expects either a Cron expression or a prefixed duration value, necessitating this change.

---------

Signed-off-by: Whit Waldo <whit.waldo@innovian.net>