Replace HashSet with ConcurrentDictionary in WorkflowLoggingService (#1556)
This should help with issue #1273 - intermittent exceptions from
WorkflowLoggingService.

There is no concurrent HashSet in .NET, so we use ConcurrentDictionary
with a byte as the value to simulate a set. For this use-case we do not
care if the Key already exists, so we can use TryAdd to add items
without considering the return value.

Signed-off-by: Tomas Ekeli <tomas@eke.li>